# Improved `GET /sca/{agent_id}/checks/{policy_id}` endpoint performance