--- a/api/test/integration/env/configurations/sca/agent/healthcheck/legacy_healthcheck.py
+++ b/api/test/integration/env/configurations/sca/agent/healthcheck/legacy_healthcheck.py
@@ -7,5 +7,5 @@ from healthcheck_utils import get_agent_health_base, check
 
 if __name__ == "__main__":
     exit(check(
-        os.system("grep -q 'sca: INFO: Security Configuration Assessment scan finished.' /var/ossec/logs/ossec.log"))
+        os.system("grep -q 'sca.*: INFO: Security Configuration Assessment scan finished.' /var/ossec/logs/ossec.log"))
          or get_agent_health_base())