… api/api/spec/spec.yaml
… api/test/integration/mapping/integration_test_api_endpoints.json
… api/test/integration/test_sca_endpoints.tavern.yaml
… framework/wazuh/core/tests/test_sca.py
… framework/wazuh/sca.py
… framework/wazuh/tests/data/schema_sca_test.sql
… framework/wazuh/tests/test_sca.py