|Related issue|
|---|
|https://github.com/wazuh/wazuh/issues/14812|


## Description

This PR closes https://github.com/wazuh/wazuh/issues/14812.

In this pull request, I have refactored the SCA framework functions used by the `GET /sca/{agent_id}/checks/{policy_id}`. After this refactoring, the responses from this API endpoint are the same but the endpoint performance is much better. This development also makes the module more flexible and new API parameters and features are now easier to implement.

I have also refactored the SDK unit tests and updated the core unit tests and API integration tests.

### Test results

```
test_sca_endpoints.tavern.yaml - Cluster environment [1/3]
	 2 passed, 3 warnings

test_sca_endpoints.tavern.yaml - Cluster environment [2/3]
	 2 passed, 3 warnings

test_sca_endpoints.tavern.yaml - Cluster environment [3/3]
	 2 passed, 3 warnings
```


### Performance changes

|                      | Before (ms) | After the refactor  (ms) |
|----------------------|-------------|------------|
| **limit=20**             | 700-800     | 25-40         |
| **limit=40**             | 700-800     | 35-45         |   
| **limit=60**             | 700-800     | 40-45         |
| **limit=80**             | 700-800     | 45-50         | 
| **limit=100**            | 700-800     | 50-60         |  
| ...            | ...     | ...         |    ...       |
| **No limit (191 items)** | 700-800     | 70-90     |